requirements (2).txt
functions-framework==3.*
requests 
google-auth